feat: refactor user services following SOLID principles
MAJOR REFACTOR - Clean Architecture Implementation:

🏗️ ARCHITECTURE CHANGES:
- Split monolithic UserService into focused, single-responsibility services
- Implement clean separation of concerns following SOLID principles
- Remove unnecessary facade pattern for direct service injection

📁 NEW STRUCTURE:
Services/User/
├── Auth/                    (Authentication Domain)
│   ├── IAuthService.cs     (Auth interface)
│   ├── AuthServiceSimple.cs (Login, Register, Password Reset)
│   └── AuthValidation.cs   (Auth validation logic)
└── Profile/                 (Profile Domain)
    ├── IProfileService.cs  (Profile interface)
    ├── ProfileServiceSimple.cs (Get, Update Profile)
    └── ProfileValidation.cs (Profile validation logic)

✅ SOLID PRINCIPLES APPLIED:
- Single Responsibility: Each service handles one domain
- Open/Closed: Easy to extend without modifying existing code
- Liskov Substitution: Services can be replaced with implementations
- Interface Segregation: Small, focused interfaces
- Dependency Inversion: Controllers depend on abstractions

🧹 CODE CLEANUP:
- Removed unnecessary XML documentation comments
- Removed verbose inline comments
- Removed region blocks and redundant explanations
- Kept only essential security and business logic comments
- Clean, professional, readable code

🔧 TECHNICAL IMPROVEMENTS:
- Direct service injection in controllers (no facade overhead)
- Simplified dependency injection setup
- Better separation of validation logic
- Maintained backward compatibility
- Enhanced maintainability and testability

🛡️ SECURITY MAINTAINED:
- All existing security validations preserved
- Profile update security checks intact
- Password hashing and upgrade logic maintained
- JWT token generation unchanged

BENEFITS:
- Easier to maintain and test individual services
- Clear separation of authentication vs profile concerns
- Reduced code complexity and file count
- Better adherence to clean architecture principles
- Improved developer experience with focused services
