--- PlatformFlower/Services/User/IUserService.cs
@@ -1,78 +1,0 @@
-using PlatformFlower.Models.DTOs;
-
-namespace PlatformFlower.Services.User
-{
-    public interface IUserService
-    {
-        /// <summary>
-        /// Register a new user and return auth response with JWT token
-        /// </summary>
-        /// <param name="registerDto">User registration data</param>
-        /// <returns>Auth response with user info and JWT token</returns>
-        Task<AuthResponseDto> RegisterUserAsync(RegisterUserDto registerDto);
-
-        /// <summary>
-        /// Login user and return auth response with JWT token
-        /// </summary>
-        /// <param name="loginDto">User login data</param>
-        /// <returns>Auth response with user info and JWT token</returns>
-        Task<AuthResponseDto> LoginUserAsync(LoginUserDto loginDto);
-
-        /// <summary>
-        /// Check if username already exists
-        /// </summary>
-        /// <param name="username">Username to check</param>
-        /// <returns>True if username exists, false otherwise</returns>
-        Task<bool> IsUsernameExistsAsync(string username);
-
-        /// <summary>
-        /// Check if email already exists
-        /// </summary>
-        /// <param name="email">Email to check</param>
-        /// <returns>True if email exists, false otherwise</returns>
-        Task<bool> IsEmailExistsAsync(string email);
-
-        /// <summary>
-        /// Get user by ID
-        /// </summary>
-        /// <param name="userId">User ID</param>
-        /// <returns>User information or null if not found</returns>
-        Task<UserResponseDto?> GetUserByIdAsync(int userId);
-
-        /// <summary>
-        /// Get user by username
-        /// </summary>
-        /// <param name="username">Username</param>
-        /// <returns>User information or null if not found</returns>
-        Task<UserResponseDto?> GetUserByUsernameAsync(string username);
-
-        /// <summary>
-        /// Send password reset email to user
-        /// </summary>
-        /// <param name="email">User email</param>
-        /// <returns>Response indicating success or failure</returns>
-        Task<ForgotPasswordResponseDto> ForgotPasswordAsync(string email);
-
-        /// <summary>
-        /// Reset user password using reset token
-        /// </summary>
-        /// <param name="resetDto">Reset password data</param>
-        /// <returns>Response indicating success or failure</returns>
-        Task<ForgotPasswordResponseDto> ResetPasswordAsync(ResetPasswordDto resetDto);
-
-        /// <summary>
-        /// Validate reset password token
-        /// </summary>
-        /// <param name="token">Reset token</param>
-        /// <returns>True if token is valid, false otherwise</returns>
-        Task<bool> ValidateResetTokenAsync(string token);
-
-        /// <summary>
-        /// Update user information including avatar upload
-        /// </summary>
-        /// <param name="userId">User ID</param>
-        /// <param name="updateDto">Update data</param>
-        /// <returns>Updated user information</returns>
-        Task<UserResponseDto> UpdateUserInfoAsync(int userId, UpdateUserInfoDto updateDto);
-    }
-}
