+++ PlatformFlower/Services/User/Profile/IProfileService.cs
@@ -1,0 +1,11 @@
+using PlatformFlower.Models.DTOs;
+
+namespace PlatformFlower.Services.User.Profile
+{
+    public interface IProfileService
+    {
+        Task<UserResponseDto?> GetUserByIdAsync(int userId);
+        Task<UserResponseDto?> GetUserByUsernameAsync(string username);
+        Task<UserResponseDto> UpdateUserInfoAsync(int userId, UpdateUserInfoDto updateDto);
+    }
+}
